Unibar, lazyblocks
Finessed the universal nav. Also noticed an odd bug where a lot of the Lazyblocks modules were getting the wrong styling. This was happening because the attributes had extra apostrophes: class="’a-class’". I noticed that in functions.php, the class attributes in the templates for the block definitions had escaped apostrophes \' instead of double quotes. Changed them all to double quotes and it works. Not sure what happened - that markup hasn't changed, so maybe WP is getting more stringent about stuff like that. (It wasn't being caused by any of the plugins I wrote; I did check that.)